soluciono problema de inconsistencia de datos en numeroCC en cuantas contables que era de tipo int y cc en base general y proveedores que era de tipo string, ahora todo es string. Tambien aseguro que la carga masiva del excel ponga CC en string, y añado el tipo a los proveedores que no estaba incluido

+++ app/Livewire/ModalNuevoCcComponent.php
@@ -28,7 +28,7 @@ class ModalNuevoCcComponent extends Component
 
         CuentaContable::create([
             // casteo el string a numero
-            $this->numeroCC = (float) $this->numeroCC,
+            $this->numeroCC = (int) $this->numeroCC,
             'numeroCC' => $this->numeroCC,
             'rubro' => $this->rubro,
             'descripcion' => strtolower($this->descripcion),
